requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
lxml==5.4.0
PyJWT==2.9.0
PyMuPDF==1.26.1
python-docx==1.2.0
sqlparse==0.5.3
typing_extensions==4.14.0
